requirements.txt
aiohttp
aiosignal
async-timeout
attrs
certifi
charset-normalizer
colorama
datasets
dill
filelock
frozenlist
fsspec
future
idna
multidict
multiprocess
numpy
packaging
pandas
Pillow
pyarrow
pyparsing
python-dateutil
pytz
PyYAML
regex
requests
responses
sentencepiece
six
tokenizers
torchvision
tqdm
transformers
typing_extensions
urllib3
xxhash
yarl